# CUDA: faster prompt processing for 4-bit quants