--- a/ggml/src/ggml-cuda/iqk_mmvq.cu
+++ b/ggml/src/ggml-cuda/iqk_mmvq.cu
@@ -409,7 +409,6 @@ __device__ __forceinline__ void vec_dot_iq4_ks_q8_1(
 
     float scale = *(const float *)vbq;
     const block_iq4_ks * bq4 = (const block_iq4_ks *)((const char *)vbq + sizeof(float)) + kbx;
-    //const uint8_t * all_values = (const uint8_t *)iq4k_values;
 
     // iqs is 0...28
     const int ib32 = iqs/4; // Why iqs/4 ?
@@ -417,22 +416,11 @@ __device__ __forceinline__ void vec_dot_iq4_ks_q8_1(
     const uint32_t * q4 = (const uint32_t *)bq4->qs + 4*ib32;
     const float dl = scale * ((bq4->scales[ib32] & 254) - 127);
     auto values = iq4k_values + ((bq4->scales[ib32] & 1) << 4);
-    //auto values = iq4k_table + ((bq4->scales[ib32] & 1) << 8);
-    //uint32_t aux32[2];
-    //auto a8 = (const uint8_t *)aux32;
-    //int v1, v2;
     int sumi = 0;
     for (int j = 0; j < 4; ++j) {
-        //aux32[0] = (q4[j] >> 0) & 0x0f0f0f0f;
-        //aux32[1] = (q4[j] >> 4) & 0x0f0f0f0f;
-        //sumi = ggml_cuda_dp4a(int_from_table_x(a8+0, values), q8[j+0], sumi);
-        //sumi = ggml_cuda_dp4a(int_from_table_x(a8+4, values), q8[j+4], sumi);
         auto v = get_int_from_table_16(q4[j], values);
         sumi = ggml_cuda_dp4a(v.x, q8[j+0], sumi);
         sumi = ggml_cuda_dp4a(v.y, q8[j+4], sumi);
-        ////get_int_from_table_16_shift(q4[j], bq4->scales[ib32] & 1, all_values, v1, v2);
-        ////sumi = ggml_cuda_dp4a(v1, q8[j+0], sumi);
-        ////sumi = ggml_cuda_dp4a(v2, q8[j+4], sumi);
     }
     *result += dl * __low2float(bq8_1[ib32].ds) * sumi;
 }
@@ -591,7 +579,6 @@ __device__ __forceinline__ void vec_dot_iq4_kss_q8_1(
 
     float scale = *(const float *)vbq;
     const block_iq4_kss * bq4 = (const block_iq4_kss *)((const char *)vbq + sizeof(float)) + kbx;
-    const uint8_t * all_values = (const uint8_t *)iq4k_values;
 
     // iqs is 0...28
     const int ib32 = iqs/4; // Why iqs/4 ?
@@ -600,14 +587,14 @@ __device__ __forceinline__ void vec_dot_iq4_kss_q8_1(
     uint32_t s32 = (q4[0] & 0x00010001) | ((q4[1] & 0x00010001) << 2) | ((q4[2] & 0x00010001) << 4) | ((q4[3] & 0x00010001) << 6);
     uint8_t ls = (s32 | (s32 >> 15)) & 0xff;
     const float dl = scale * ((ls & 254) - 127);
-    int v1, v2;
+    auto values = iq4k_values + ((ls & 1) << 4);
     int sumi = 0;
     for (int j = 0; j < 4; ++j) {
         uint32_t aux32 = q4[j] & 0xfffefffe;
         aux32 ^= (aux32 >> 1);
-        get_int_from_table_16_shift(aux32, ls & 1, all_values, v1, v2);
-        sumi = ggml_cuda_dp4a(v1, q8[j+0], sumi);
-        sumi = ggml_cuda_dp4a(v2, q8[j+4], sumi);
+        auto v = get_int_from_table_16(aux32, values);
+        sumi = ggml_cuda_dp4a(v.x, q8[j+0], sumi);
+        sumi = ggml_cuda_dp4a(v.y, q8[j+4], sumi);
     }
     *result += dl * __low2float(bq8_1[ib32].ds) * sumi;
 }