--- a/ggml/src/ggml-cuda/template-instances/mmq-instance-iq4_kss.cu
+++ b/ggml/src/ggml-cuda/template-instances/mmq-instance-iq4_kss.cu
@@ -14,9 +14,6 @@ template <int mmq_y, int nwarps, bool need_check> static __device__ __forceinlin
 
     const int kqsx = threadIdx.x / 4;
 
-    uint32_t aux32[2];
-    auto a8 = (const uint8_t *)aux32;
-
 #pragma unroll
     for (int i0 = 0; i0 < mmq_y; i0 += 4*nwarps) {
         int i = i0 + 4*threadIdx.y + threadIdx.x%4;
@@ -31,20 +28,19 @@ template <int mmq_y, int nwarps, bool need_check> static __device__ __forceinlin
         uint32_t s32 = (q4[0] & 0x00010001) | ((q4[1] & 0x00010001) << 2) | ((q4[2] & 0x00010001) << 4) | ((q4[3] & 0x00010001) << 6);
         uint8_t ls = (s32 | (s32 >> 15)) & 0xff;
 
-        auto values = iq4k_table + ((ls & 1) << 8);
+        auto values = iq4k_values + ((ls & 1) << 4);
 
         #pragma unroll
         for (int j = 0; j < 4; ++j) {
             uint32_t val = q4[j] & 0xfffefffe;
             val = val ^ (val >> 1);
-            aux32[0] = (val >> 0) & 0x0f0f0f0f;
-            aux32[1] = (val >> 4) & 0x0f0f0f0f;
+            auto v = get_int_from_table_16(val, values);
 #ifdef INT8_MMA_AVAILABLE
-            x_qs[i*MMQ_MMA_TILE_X_K_Q8_0 + 8*kqsx + j + 0] = int_from_table_x(a8+0, values);
-            x_qs[i*MMQ_MMA_TILE_X_K_Q8_0 + 8*kqsx + j + 4] = int_from_table_x(a8+4, values);
+            x_qs[i*MMQ_MMA_TILE_X_K_Q8_0 + 8*kqsx + j + 0] = v.x;
+            x_qs[i*MMQ_MMA_TILE_X_K_Q8_0 + 8*kqsx + j + 4] = v.y;
 #else
-            x_qs[i*(2*WARP_SIZE + 1)     + 8*kqsx + j + 0] = int_from_table_x(a8+0, values);
-            x_qs[i*(2*WARP_SIZE + 1)     + 8*kqsx + j + 4] = int_from_table_x(a8+4, values);
+            x_qs[i*(2*WARP_SIZE + 1)     + 8*kqsx + j + 0] = v.x;
+            x_qs[i*(2*WARP_SIZE + 1)     + 8*kqsx + j + 4] = v.y;
 #endif // INT8_MMA_AVAILABLE
         }
 #ifdef INT8_MMA_AVAILABLE