@Ph0rk0z 2025-08-21
A hair under 20t/s up from 18 on qwen 235b.. wooo!